requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
shapely<2.0.0
google-cloud-aiplatform>=1.26.0
arxiv
arxiv2bib
pandas
regex
vertexai
subprocess.run
nltk
typing
